update name of Editor component
