+++ src/common/components/PostInvocation/TestsEditor.tsx
@@ -1,0 +1,45 @@
+import { Editor as MonacoEditor } from '@monaco-editor/react';
+import { Dispatch, SetStateAction } from 'react';
+
+import EditorHelpers from '../PreInvocate/EditorHelpers';
+
+import {
+	editorTestsHelpers,
+	editorTestsHelpersDescriptions,
+} from '@/common/constants/editor-helpers';
+
+function PostInvocationEditor({
+	postInvocationValue,
+	setPostInvocationValue,
+}: {
+	postInvocationValue: string;
+	setPostInvocationValue: Dispatch<SetStateAction<string>>;
+}) {
+	function handleEditorChange(value: string | undefined) {
+		setPostInvocationValue(value ?? '');
+	}
+
+	return (
+		<div className="flex h-full">
+			<MonacoEditor
+				data-test="editor"
+				height={'100%'}
+				width={'80%'}
+				defaultLanguage="javascript"
+				onChange={handleEditorChange}
+				theme="vs-dark"
+				defaultValue={postInvocationValue}
+				value={postInvocationValue}
+			/>
+			<EditorHelpers
+				setEditorValue={setPostInvocationValue}
+				title="	Post invocations are written in Javascript, and are run after the
+				invocation."
+				editorHelpers={editorTestsHelpers}
+				editorHelpersDescriptions={editorTestsHelpersDescriptions}
+			/>
+		</div>
+	);
+}
+
+export default PostInvocationEditor;
